Modified the way the remote repository is found.
Instead of just trying to find a (remote) branch called trunk and
assuming we are tracking that branch, acually do some factual checking
to find a remote branch.

When working on e.g. a local branch based on a branch in the
subversion repository, this will result in a more accurate listing of
unpushed commits.